requirements.txt
langchain-community
langchain-openai
langchain
chromadb
python-dotenv
openai
tiktoken